+++ docs/history.rst
@@ -3,8 +3,8 @@
 Release History
 ===============
 
-`Next Release`_
----------------
+`1.0.0`_ (20 Nov 2015)
+----------------------
 - Add ``sprockets.http.mixins.LoggingHandler``
 - Add ``sprockets.http.mixins.ErrorLogger``
 - Add ``sprockets.http.mixins.ErrorWriter``
@@ -37,4 +37,5 @@ Release History
 .. _0.2.2: https://github.com/sprockets/sprockets.http/compare/0.2.1...0.2.2
 .. _0.3.0: https://github.com/sprockets/sprockets.http/compare/0.2.2...0.3.0
 .. _0.4.0: https://github.com/sprockets/sprockets.http/compare/0.3.0...0.4.0
-.. _Next Release: https://github.com/sprockets/sprockets.http/compare/0.4.0...master
+.. _1.0.0: https://github.com/sprockets/sprockets.http/compare/0.4.0...1.0.0
+.. _Next Release: https://github.com/sprockets/sprockets.http/compare/1.0.0...master